Log user out from UI on 401
