+++ ui/src/api/api.js
@@ -1,0 +1,16 @@
+/* eslint-disable promise/prefer-await-to-callbacks */
+import baseApi from './apiBase';
+import { fetchLoginInfo } from './loginBase';
+import { store } from '../store';
+
+const api = baseApi
+  .options({ credendials: 'include' })
+  .catcher(401, async (error) => {
+    // User got logged out somehow: refetch login info to update local state and redirect to login page.
+    // Don't use `getLoginInfo` action to avoid import cycle.
+    const loginInfo = await fetchLoginInfo();
+    store.dispatch({ type: 'SET_LOGIN_INFO', loginInfo });
+    throw error;
+  });
+
+export default api;
